cinii-mcp-server/requirements.txt
fastapi
httpx
uvicorn
mcp-server==0.1.0
pydantic
